requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
astroid==3.3.5
attrs==22.1.0
auto_subtitle @ git+https://github.com/m1guelpf/auto-subtitle.git@124ccb1ac17d5b7a27dd81b3d8f8fed6ef1a5408
autopep8==2.3.2
bandit==1.8.3
black==25.1.0
blis==1.2.0
boto==2.49.0
boto3==1.36.26
botocore==1.36.26
branca==0.8.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
coverage==7.10.5
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
dill==0.3.9
distlib==0.3.9
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.17.0
flake8==7.3.0
folium==0.19.5
fonttools==4.59.1
fsspec==2024.12.0
future==1.0.0
geopandas==1.0.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.1
kagglehub==0.3.8
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
moto==5.0.28
mpmath==1.3.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pbr==6.1.0
pdfkit==1.0.0
pg8000==1.31.2
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
preshed==3.0.9
psutil==7.1.0
py-partiql-parser==0.6.1
pyarrow==19.0.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.2
pyodbc==5.2.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
PySocks==1.7.1
pytest==8.4.1
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.25.6
rich==13.9.4
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.7.1
scipy==1.15.3
scramp==1.4.5
seaborn==0.13.2
selenium==4.27.1
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.47.1
stevedore==5.4.0
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.29.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
xyzservices==2025.1.0
